exercise_6_langgraph_langsmith/pyproject.toml
[project]
name = "exercise-6-langgraph-langsmith"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.10",
    "langchain-groq>=1.1.2",
    "langgraph>=1.0.9",
    "langsmith>=0.7.6",
    "notebook>=7.5.4",
    "python-dotenv>=1.2.1",
]
